@@ -18,7 +18,7 @@ import '../../utils/toast.dart';
 import '../../widgets/common/meeting_controls/meeting_action_bar.dart';
 import '../../widgets/common/screen_share/screen_select_dialog.dart';
 import '../common/join_screen.dart';
-import 'package:flutter_webrtc/flutter_webrtc.dart';
+import 'package:videosdk_webrtc/flutter_webrtc.dart';
 
 // Meeting Screen
 class OneToOneMeetingScreen extends StatefulWidget {
@@ -26,6 +26,7 @@ class OneToOneMeetingScreen extends StatefulWidget {
   final bool micEnabled, camEnabled, chatEnabled;
   final AudioDeviceInfo? selectedAudioOutputDevice, selectedAudioInputDevice;
   final CustomTrack? cameraTrack;
+  final CustomTrack? micTrack;
 
   const OneToOneMeetingScreen(
       {Key? key,
@@ -37,7 +38,8 @@ class OneToOneMeetingScreen extends StatefulWidget {
       this.chatEnabled = true,
       this.selectedAudioOutputDevice,
       this.selectedAudioInputDevice,
-      this.cameraTrack})
+      this.cameraTrack,
+      this.micTrack})
       : super(key: key);
 
   @override
@@ -78,7 +80,7 @@ class _OneToOneMeetingScreenState extends State<OneToOneMeetingScreen> {
 
     // Create instance of Room (Meeting)
     Room room = VideoSDK.createRoom(
-      roomId: widget.meetingId,
+      roomId: "oupf-fhti-93e9",
       token: widget.token,
       displayName: widget.displayName,
       micEnabled: widget.micEnabled,
@@ -86,11 +88,7 @@ class _OneToOneMeetingScreenState extends State<OneToOneMeetingScreen> {
       maxResolution: 'hd',
       multiStream: false,
       customCameraVideoTrack: widget.cameraTrack,
-      // defaultCameraIndex: kIsWeb
-      //     ? 0
-      //     : (Platform.isAndroid || Platform.isIOS)
-      //         ? 1
-      //         : 0,
+      customMicrophoneAudioTrack: widget.micTrack,
       notification: const NotificationInfo(
         title: "Video SDK",
         message: "Video SDK is sharing screen in the meeting",
@@ -112,8 +110,13 @@ class _OneToOneMeetingScreenState extends State<OneToOneMeetingScreen> {
     bool isWebMobile = kIsWeb &&
         (defaultTargetPlatform == TargetPlatform.iOS ||
             defaultTargetPlatform == TargetPlatform.android);
-    return WillPopScope(
-      onWillPop: _onWillPopScope,
+    return PopScope(
+      onPopInvoked: (didPop) async {
+        if (didPop) {
+          return;
+        }
+        _onWillPopScope();
+      },
       child: _joined
           ? SafeArea(
               child: Scaffold(
@@ -192,10 +195,10 @@ class _OneToOneMeetingScreenState extends State<OneToOneMeetingScreen> {
                                     },
 
                                     onSwitchMicButtonPressed: (details) async {
-                                      List<MediaDeviceInfo> outptuDevice =
-                                          meeting.getAudioOutputDevices();
+                                      List<AudioDeviceInfo>? outputDevice =
+                                          await VideoSDK.getAudioDevices();
                                       double bottomMargin =
-                                          (70.0 * outptuDevice.length);
+                                          (70.0 * outputDevice!.length);
                                       final screenSize =
                                           MediaQuery.of(context).size;
                                       await showMenu(
@@ -212,7 +215,7 @@ class _OneToOneMeetingScreenState extends State<OneToOneMeetingScreen> {
                                           details.globalPosition.dx,
                                           (bottomMargin),
                                         ),
-                                        items: outptuDevice.map((e) {
+                                        items: outputDevice.map((e) {
                                           return PopupMenuItem(
                                               value: e, child: Text(e.label));
                                         }).toList(),
@@ -324,11 +327,9 @@ class _OneToOneMeetingScreenState extends State<OneToOneMeetingScreen> {
             meeting = _meeting;
             _joined = true;
           });
-          _meeting.switchAudioDevice(
-              widget.selectedAudioOutputDevice!);
+
           if (kIsWeb || Platform.isWindows || Platform.isMacOS) {
-            _meeting
-                .changeMic(widget.selectedAudioInputDevice!);
+            _meeting.switchAudioDevice(widget.selectedAudioOutputDevice!);
           }
           subscribeToChatMessages(_meeting);
         }
